ml-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0

# Audio
librosa>=0.10.2
soundfile>=0.13.0
noisereduce>=3.0.0

# Acoustic analysis (Praat)
praat-parselmouth>=0.4.5

# Speech-to-text
openai-whisper>=20240930

# ML
scikit-learn>=1.6.0
xgboost>=2.1.0
joblib>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Storage
minio>=7.2.0
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0
